+++ podverse/FindTableViewController.swift
@@ -48,16 +48,6 @@ class FindTableViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
 
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-        if segue.identifier == Constants.TO_PLAYER_SEGUE_ID {
-            let mediaPlayerViewController = segue.destinationViewController as! MediaPlayerViewController
-            mediaPlayerViewController.hidesBottomBarWhenPushed = true
-        }
-    }
-
 }
 
 extension FindTableViewController:UITableViewDelegate, UITableViewDataSource {
@@ -114,6 +104,16 @@ extension FindTableViewController:UITableViewDelegate, UITableViewDataSource {
             }
         }
         tableView.deselectRowAtIndexPath(indexPath, animated: true)
+    }
+    
+    // MARK: - Navigation
+    
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        if segue.identifier == Constants.TO_PLAYER_SEGUE_ID {
+            let mediaPlayerViewController = segue.destinationViewController as! MediaPlayerViewController
+            mediaPlayerViewController.hidesBottomBarWhenPushed = true
+        }
     }
     
 }
