+++ podverse/podverse.xcdatamodeld/podverse.xcdatamodel/contents
@@ -42,6 +42,7 @@
         <attribute name="isMyEpisodes" optional="YES" attributeType="Boolean" syncable="YES"/>
         <attribute name="lastUpdated" optional="YES" attributeType="Date" syncable="YES"/>
         <attribute name="ownerId" optional="YES" attributeType="String" syncable="YES"/>
+        <attribute name="ownerName" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="podverseURL" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="sharePermission" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="title" optional="YES" attributeType="String" syncable="YES"/>
@@ -66,7 +67,7 @@
     <elements>
         <element name="Clip" positionX="-20" positionY="-387" width="128" height="210"/>
         <element name="Episode" positionX="-225" positionY="-675" width="128" height="345"/>
-        <element name="Playlist" positionX="-288" positionY="-265" width="128" height="210"/>
+        <element name="Playlist" positionX="-288" positionY="-265" width="128" height="225"/>
         <element name="Podcast" positionX="-398" positionY="-541" width="128" height="240"/>
     </elements>
 </model>
